+++ SourceCode/Data/LayoutModule.cs
@@ -2,6 +2,7 @@
 
 using Microsoft.AspNetCore.Components;
 using Microsoft.EntityFrameworkCore;
+using System.Text;
 
 namespace ModulesRegistry.Data;
 
@@ -29,7 +30,7 @@ public static class LayoutModuleExtensions
 {
     public static MarkupString Info(this LayoutModule it, LayoutStation? station)
     {
-        if (station is not null) 
+        if (station is not null)
             it.Module.Station = it.LayoutStation.Station;
         return it.Module.Description();
     }
@@ -38,12 +39,26 @@ public static class LayoutModuleExtensions
     public static bool HasCargoCustomers(this LayoutModule it) => it.LayoutStation?.Station is not null && it.LayoutStation.Station.HasCargoCustomers;
     public static bool HasDwgDrawing(this LayoutModule it) => it.Module.DwgDrawing.Id > 0;
     public static bool HasSkpDrawing(this LayoutModule it) => it.Module.SkpDrawing.Id > 0;
+    public static bool IsBorrowed(this LayoutModule it) => !it.Module.ModuleOwnerships.Any(mo => mo.GroupId > 0 || mo.PersonId == it.LayoutParticipant.PersonId);
+
     public static double TotalLength(this LayoutModule it) => it.Module?.TotalLength() ?? 0.0;
     public static double TotalLengthMeters(this IEnumerable<LayoutModule> modules)
     {
         double sum = 0.0;
         foreach (var module in modules) sum += module.TotalLength();
-        return sum/1000;
+        return sum / 1000;
+    }
+    public static string FullNameAndConfiguration(this LayoutModule it)
+    {
+        var result = new StringBuilder(100);
+        result.Append(it.Module.FullNameAndConfiguration());
+        if (it.IsBorrowed())
+        {
+            result.Append(" (");
+            result.Append(string.Format(Resources.Strings.BorrowedFrom.ToLowerInvariant(),it.Module.ModuleOwnerships.OwnerNames() ));
+            result.Append(')');
+        }
+        return result.ToString();
     }
 }